+++ packages/orq-rc/RELEASES.md
@@ -169,3 +169,13 @@ Based on:
 - [typescript v3.1.0-rc.37] packages/orq-rc
 ### Releases
 - [NPM v3.1.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.37 - packages/orq-rc
+
+## 2025-02-05 18:58:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.484.1 (2.503.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.1.0-rc.38] packages/orq-rc
+### Releases
+- [NPM v3.1.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.38 - packages/orq-rc